Update makefile to add a `build_rules`
It would be really useful to be able to pull in `core` to
a repo the same way `KE-complex_modifications` does. It's entirely
possible to make rules that are not useful or reasonable to have
polluting the public page.

This allows one to do that and run `make -C core build_rules`, which
will just build the rules without trying to build the 'groups' or
webpage stuff.

Currently this is how I have my repo setup, and then I simply symlink
`~/.config/karabiner/assets/<file>` to `<repo>/public/json/<file>`.

This modification has `all` call `build_rules`, so there's absolutely
no change in workflow for anyone contributing their rules to the
primary repo.

Signed-off-by: Phil Dibowitz <phil@ipom.com>